Fixed incorrect autocomple of hyphenated classes.
Function returning current word entry range was added, which  is used to identify the range of the class name being currently entered by the user, including hyphens.

The problem it solves is that an autocomplete will insert the entire label and will not replace any of class name before a hyphen. For example: If a user types “bg-re” and then uses the autocomplete to select “bg-red-100”, the inserted class will be “bg-bg-red-100”.